+++ src/main/java/com/social_media/social_media/controller/UserController.java
@@ -1,11 +1,10 @@
 package com.social_media.social_media.controller;
 
-import com.social_media.social_media.dto.responseDto.FollowingResponseDto;
-import com.social_media.social_media.service.user.IUserService;
+import com.social_media.social_media.service.IUserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @RequiredArgsConstructor
 @RestController
@@ -17,8 +16,4 @@ public class UserController {
         return new String("Hola Mundo");
     }
 
-    @PostMapping("/{userId}/follow/{userIdToFollow}")
-    public ResponseEntity<FollowingResponseDto> postFollowSeller(@PathVariable Long userId, @PathVariable Long userIdToFollow){
-        return new ResponseEntity<>(userService.followSeller(userId, userIdToFollow), HttpStatus.OK);
-    }
 }
