--- src/main/java/com/social_media/social_media/exception/BadRequestFollowException.java
@@ -1,7 +1,0 @@
-package com.social_media.social_media.exception;
-
-public class BadRequestFollowException extends RuntimeException {
-    public BadRequestFollowException(String message) {
-        super(message);
-    }
-}
