+++ src/main/java/com/social_media/social_media/repository/follow/IFollowRepository.java
@@ -1,11 +1,5 @@
 package com.social_media.social_media.repository.follow;
 
-import com.social_media.social_media.dto.responseDto.FollowingResponseDto;
-import com.social_media.social_media.entity.Follow;
-
 public interface IFollowRepository {
 
-    Follow addFollow(Long followerId, Long followedId);
-
-    boolean existsByFollowerAndFollowed(Long userId, Long userIdToFollow);
 }
