+++ src/main/java/com/social_media/social_media/service/IUserService.java
@@ -1,0 +1,4 @@
+package com.social_media.social_media.service;
+
+public interface IUserService {
+}
