--- src/main/java/com/social_media/social_media/service/post/IPostService.java
@@ -1,4 +1,0 @@
-package com.social_media.social_media.service.post;
-
-public interface IPostService {
-}
