--- src/main/java/com/social_media/social_media/service/user/IUserService.java
@@ -1,8 +1,0 @@
-package com.social_media.social_media.service.user;
-
-import com.social_media.social_media.dto.responseDto.FollowingResponseDto;
-
-public interface IUserService {
-
-    FollowingResponseDto followSeller(Long userId, Long userIdToFollow);
-}
