+++ src/main/java/com/social_media/social_media/utils/MessagesExceptions.java
@@ -7,6 +7,4 @@ public final class MessagesExceptions {
     public static final String SELLER_ID_NOT_EXIST = "El vendedor no existe";
     public static final String BUYER_ID_NOT_EXIST = "El comprador no existe";
     public static final String FOLLOWED_USER_NOT_SELLER = "El usuario a seguir no es un vendedor";
-    public static final String THE_USER_CANNOT_FOLLOW_THEMSELVES = "El usuario no se puede seguir a si mismo";
-    public static final String FOLLOW_ALREADY_EXISTS = "Este usuario ya está siguiendo al usuario especificado.";
 }
